fix: SSRF 취약점 수정 — HTTP 자기호출을 직접 함수 호출로 전환
req.nextUrl.origin이 Host 헤더에 의존하여 SSRF 가능성 있음.
processUploadQueue 함수를 분리하여 네트워크 요청 없이 직접 호출.
